Restyled Create post/comment screen
